When canceling a visit, navigation must be canceled before the request
Canceling navigation cancels the request, but canceling a request doesn't
cancel navigation. This makes sense, since it's not possible to cancel a
request while allowing navigation to go through.

This ensures that navigation completion actions aren't run when finish()
calls completeNavigation(), since by this time, navigation will already be
canceled.

Fixes that canceled visits in the backward direction would result in
navigation completion, causing the web view to be incorrectly activated on
the current visitable.