Update the screenshot before we cancel the visit
This ensures that if we're already showing a screenshot, nothing happens.
After we cancel we won't be showing a screenshot, and updateScreenshot()
will therefore capture a new one. If no snapshot was restored, the view
won't have yet redrawn, and our activity indicator will be captured along
with it.

If we *have* restored a snapshot, we won't be showing a screenshot or an
activity indicator. A fresh screenshot of the restored page will be taken.