Wait until ColdBootVisit completes successfully before calling visitDidInitializeWebView
A ColdBootVisit completes successfully when it receives
WebViewPageLoadDelegate's webView(_:didLoadPageWithRestorationIdentifier)
message. This message originates from web view's Turbolinks adapter, implying
that Turbolinks was successfully loaded and can communicate over the JavaScript
bridge.

Until we've received this message, we can't be sure Turbolinks is installed and
working, and shouldn't consider the web view initialized.