Make sure finishRequest() is always called
When doing a cold boot visit, `finishRequest()` was previously only
called in response to receiving the `webView(_:didFinish:)`
`WKNavigationDelegate` message. If the web view wasn't on screen when
the navigation finished—for example after the user switched to another
tab in a parent `UITabViewController`—`webView(_:didFinish:)` would not
be called, causing the `VisitDelegate` to never be sent the
`visitRequestDidFinish(_:)` message.

Since `finishRequest()` will only call `visitRequestDidFinish(_:)` once,
it should be safe to call it again in `completeVisit()` at which point
the request must have finished since the whole visit has already
completed.