Update Visitable.visitableURL after following a redirect
Previously, after visiting a URL that redirected to another URL,
pull-to-refresh would reload the original URL not the one that was
redirected to.

This is inconsistent with how Turbolinks works in a browser where the
address bar is updated using the History API causing reloading the page
to reload the URL that was redirected to.